Added an option to use local times instead of universal times and duplicated the first track point as second 0 to make the times come out on even minutes when 60 seconds are captured in the file.
The option to use local time allows strava.com to see the ride on the correct day.
Strava.com has a problem because TCX files are always in UTC and there is no GPS data from
a trainer, they don't know from the GPS data what local time to convert the UTC to.  That means
that if you are PST (-8) and you start a ride past 4pm it will look like is was
on the following day because it is past midnight in UTC time.

duplicating the first point as second 0 allows the times to come out the same duration that was riden.
When riding for exatly 1 minute the .CSV file will have 60 data lines, and the created .TCX will also,
but strava.com sees that as a 59 second ride.